requirements.txt
python_libsbml
python_libsedml
python_libcombine
beautifulsoup4
requests
lxml
docker
pkg_utils
pandas
logzero
